feat: svg 파일 분리
