+++ src/widgets/homeMeisterSection/ui/ChevronDownIcon.tsx
@@ -1,0 +1,10 @@
+const ChevronDownIcon = () => (
+  <svg width="12" height="14" viewBox="0 0 12 14" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path
+      d="M6.00197 11.9638L10.5734 7.23192C10.7288 7.07047 10.9212 6.98489 11.1507 6.97517C11.3802 6.96545 11.582 7.05103 11.7561 7.23192C11.9187 7.40086 12 7.60568 12 7.84639C12 8.08708 11.9187 8.2919 11.7561 8.46083L6.71205 13.7017C6.60702 13.8109 6.49622 13.8879 6.37967 13.9327C6.26313 13.9776 6.13723 14 6.00197 14C5.86671 14 5.74081 13.9776 5.62428 13.9327C5.50772 13.8879 5.39693 13.8109 5.2919 13.7017L0.247855 8.46083C0.092448 8.29938 0.010068 8.09942 0.000715191 7.86097C-0.00863763 7.62251 0.0737424 7.41283 0.247855 7.23192C0.410444 7.06299 0.607564 6.97852 0.839214 6.97852C1.07087 6.97852 1.26798 7.06299 1.43057 7.23192L6.00197 11.9638ZM6.00197 4.98938L10.5734 0.257527C10.7288 0.0960552 10.9212 0.0104609 11.1507 0.000743109C11.3802 -0.00897473 11.582 0.0766195 11.7561 0.257527C11.9187 0.426462 12 0.631275 12 0.871966C12 1.11266 11.9187 1.31747 11.7561 1.48641L6.71205 6.7273C6.60702 6.83646 6.49622 6.91345 6.37967 6.95829C6.26313 7.00315 6.13723 7.02557 6.00197 7.02557C5.86671 7.02557 5.74081 7.00315 5.62428 6.95829C5.50772 6.91345 5.39693 6.83646 5.2919 6.7273L0.247855 1.48641C0.092448 1.32495 0.010068 1.125 0.000715191 0.886543C-0.00863763 0.648086 0.0737424 0.438415 0.247855 0.257527C0.410444 0.0885725 0.607564 0.00409548 0.839214 0.00409548C1.07087 0.00409548 1.26798 0.0885725 1.43057 0.257527L6.00197 4.98938Z"
+      fill="white"
+    />
+  </svg>
+);
+
+export default ChevronDownIcon;
