+++ src/widgets/homeMeisterSection/ui/HomeMeisterSection.tsx
@@ -1,104 +1,5 @@
-const ChevronDownIcon = () => (
-  <svg width="12" height="14" viewBox="0 0 12 14" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <path
-      d="M6.00197 11.9638L10.5734 7.23192C10.7288 7.07047 10.9212 6.98489 11.1507 6.97517C11.3802 6.96545 11.582 7.05103 11.7561 7.23192C11.9187 7.40086 12 7.60568 12 7.84639C12 8.08708 11.9187 8.2919 11.7561 8.46083L6.71205 13.7017C6.60702 13.8109 6.49622 13.8879 6.37967 13.9327C6.26313 13.9776 6.13723 14 6.00197 14C5.86671 14 5.74081 13.9776 5.62428 13.9327C5.50772 13.8879 5.39693 13.8109 5.2919 13.7017L0.247855 8.46083C0.092448 8.29938 0.010068 8.09942 0.000715191 7.86097C-0.00863763 7.62251 0.0737424 7.41283 0.247855 7.23192C0.410444 7.06299 0.607564 6.97852 0.839214 6.97852C1.07087 6.97852 1.26798 7.06299 1.43057 7.23192L6.00197 11.9638ZM6.00197 4.98938L10.5734 0.257527C10.7288 0.0960552 10.9212 0.0104609 11.1507 0.000743109C11.3802 -0.00897473 11.582 0.0766195 11.7561 0.257527C11.9187 0.426462 12 0.631275 12 0.871966C12 1.11266 11.9187 1.31747 11.7561 1.48641L6.71205 6.7273C6.60702 6.83646 6.49622 6.91345 6.37967 6.95829C6.26313 7.00315 6.13723 7.02557 6.00197 7.02557C5.86671 7.02557 5.74081 7.00315 5.62428 6.95829C5.50772 6.91345 5.39693 6.83646 5.2919 6.7273L0.247855 1.48641C0.092448 1.32495 0.010068 1.125 0.000715191 0.886543C-0.00863763 0.648086 0.0737424 0.438415 0.247855 0.257527C0.410444 0.0885725 0.607564 0.00409548 0.839214 0.00409548C1.07087 0.00409548 1.26798 0.0885725 1.43057 0.257527L6.00197 4.98938Z"
-      fill="white"
-    />
-  </svg>
-);
-
-const GSM = () => (
-  <svg xmlns="http://www.w3.org/2000/svg" width="295" height="24" viewBox="0 0 295 24" fill="none">
-    <path
-      d="M2.43594 0C3.36839 0 18.7539 0 19.6863 0C20.6188 0 22.0175 1.28693 22.0175 2.14488C22.0175 3.00283 22.0175 5.3817 22.0175 5.3817H0.104802C0.104802 5.3817 0.104802 3.00283 0.104802 2.14488C0.104802 1.28693 1.50348 0 2.43594 0Z"
-      fill="white"
-    />
-    <path
-      d="M26.6798 0C27.6123 0 42.9978 0 43.9302 0C44.8627 0 46.2613 1.28693 46.2613 2.14488C46.2613 3.00283 46.2613 5.3817 46.2613 5.3817H24.3487C24.3487 5.3817 24.3487 3.00283 24.3487 2.14488C24.3487 1.28693 25.7474 0 26.6798 0Z"
-      fill="white"
-    />
-    <path
-      d="M54.1872 0L58.1501 5.3817H48.5924C48.5924 5.3817 48.5924 3.00283 48.5924 2.14488C48.5924 1.28693 49.8919 0 50.9236 0H54.1872Z"
-      fill="white"
-    />
-    <path
-      d="M70.5051 0C71.5367 0 72.8362 1.28693 72.8362 2.14488V5.3817H63.2786L67.2415 0H70.5051Z"
-      fill="white"
-    />
-    <path
-      d="M0 7.15295H6.06095V17.9163H15.8517V14.6405H13.0543L9.32453 9.96079H21.9127C21.9127 9.96079 21.9127 20.7242 21.9127 21.6601C21.9127 22.5961 20.514 24 19.5815 24C18.6491 24 3.26359 24 2.33113 24C1.39868 24 0 22.5961 0 21.6601C0 20.7242 0 7.15295 0 7.15295Z"
-      fill="white"
-    />
-    <path
-      d="M46.2613 21.5264C46.2613 22.4624 44.8627 23.8663 43.9302 23.8663H26.6798C25.7474 23.8663 24.3487 22.4624 24.3487 21.5264V17.7827H37.6361L24.3487 7.01926H32.7408L46.2613 17.7827V21.5264Z"
-      fill="white"
-    />
-    <path
-      d="M54.1872 9.35913V23.3983H48.5924V7.01926H58.8494L60.7143 9.8271L62.5792 7.01926H72.8362C72.8362 7.01926 72.8362 20.1225 72.8362 21.0585C72.8362 21.9944 71.4375 23.3983 70.5051 23.3983C69.5726 23.3983 67.2415 23.3983 67.2415 23.3983V9.35913L60.7143 18.2506L54.1872 9.35913Z"
-      fill="white"
-    />
-    <path
-      d="M90.8915 5.2999V6.23983C90.8755 7.17975 90.8755 8.47016 90.5382 10.0473L88.5305 9.90394C88.7634 8.68523 88.8196 7.73733 88.8357 6.92486H83.3427V5.2999H90.8915ZM87.2295 8.53388V11.035C88.6911 10.9872 90.1768 10.8917 91.5501 10.7324L91.6625 12.1502C88.6269 12.6759 85.2861 12.7556 82.652 12.7556L82.4754 11.1306C83.3106 11.1227 84.2261 11.1227 85.1897 11.0988V8.53388H87.2295ZM94.4411 4.50335V8.07188H96.2239V9.76056H94.4411V13.345H92.3371V4.50335H94.4411ZM89.4621 13.5203C92.5941 13.5203 94.5375 14.5558 94.5375 16.2126C94.5375 17.965 92.5941 18.9687 89.4621 18.9687C86.3301 18.9687 84.3706 17.965 84.3706 16.2126C84.3706 14.5558 86.3301 13.5203 89.4621 13.5203ZM89.4621 15.0815C87.5347 15.0815 86.4586 15.4798 86.4586 16.2126C86.4586 17.041 87.5347 17.3915 89.4621 17.3915C91.3734 17.3915 92.4334 17.041 92.4334 16.2126C92.4334 15.4798 91.3734 15.0815 89.4621 15.0815Z"
-      fill="white"
-    />
-    <path
-      d="M110.213 12.5007V14.1256H104.528V19.0005H102.504V14.1256H96.8825V12.5007H110.213ZM109.025 5.26804V6.86113H104.769C104.969 8.27899 106.584 9.70481 109.587 10.0633L108.848 11.6563C106.278 11.3138 104.439 10.2226 103.564 8.72505C102.665 10.2226 100.842 11.3138 98.2959 11.6563L97.541 10.0633C100.496 9.70481 102.127 8.27899 102.359 6.86113H98.0871V5.26804H109.025Z"
-      fill="white"
-    />
-    <path
-      d="M124.476 15.6391V17.264H111.097V15.6391H116.686V12.3573H118.726V15.6391H124.476ZM118.79 6.36727C118.79 8.4383 120.669 10.6208 124.01 11.1306L123.143 12.8193C120.573 12.3573 118.662 10.9952 117.698 9.18705C116.734 10.9872 114.839 12.3653 112.285 12.8193L111.418 11.1306C114.727 10.6208 116.59 8.42237 116.606 6.36727V5.33176H118.79V6.36727Z"
-      fill="white"
-    />
-    <path
-      d="M138.69 15.6232V17.3118H125.311V15.6232H138.69ZM137.566 5.77783V7.40278H135.51V11.7997H137.502V13.3928H126.419V11.7997H128.379V7.40278H126.355V5.77783H137.566ZM130.435 11.7997H133.486V7.40278H130.435V11.7997Z"
-      fill="white"
-    />
-    <path
-      d="M151.443 11.9909V13.5999H141.1V5.57073H151.331V7.21161H143.188V8.74098H150.993V10.3341H143.188V11.9909H151.443ZM152.905 15.655V17.3118H139.526V15.655H152.905Z"
-      fill="white"
-    />
-    <path
-      d="M157.418 5.31583C159.281 5.31583 160.646 6.46286 160.662 8.07188C160.646 9.68091 159.281 10.812 157.418 10.812C155.523 10.812 154.141 9.68091 154.141 8.07188C154.141 6.46286 155.523 5.31583 157.418 5.31583ZM157.418 6.81334C156.535 6.81334 155.956 7.29127 155.956 8.07188C155.956 8.8525 156.535 9.29857 157.418 9.3145C158.253 9.29857 158.831 8.8525 158.847 8.07188C158.831 7.29127 158.253 6.81334 157.418 6.81334ZM166.589 4.50335V19.0005H164.662V4.50335H166.589ZM163.634 4.74232V18.3314H161.706V15.3683H159.377V13.823H161.706V4.74232H163.634ZM153.531 11.5767C155.491 11.5926 158.655 11.5767 160.967 11.2262L161.128 12.7397C160.253 12.9069 159.289 13.0184 158.333 13.0901V17.8535H156.406V13.2016C155.426 13.2335 154.503 13.2415 153.756 13.2335L153.531 11.5767Z"
-      fill="white"
-    />
-    <path
-      d="M171.986 5.52293C173.961 5.52293 175.431 7.13992 175.728 9.74463H178.266V4.50335H180.289V19.0324H178.266V11.3696H175.744C175.527 14.1575 174.025 15.8621 171.986 15.8621C169.769 15.8621 168.179 13.8707 168.195 10.6846C168.179 7.53023 169.769 5.52293 171.986 5.52293ZM171.986 7.35499C170.893 7.35499 170.139 8.53388 170.155 10.6846C170.139 12.8512 170.893 14.0141 171.986 14.0141C173.094 14.0141 173.817 12.8512 173.817 10.6846C173.817 8.53388 173.094 7.35499 171.986 7.35499Z"
-      fill="white"
-    />
-    <path
-      d="M189.653 5.87341V15.4638H182.634V5.87341H189.653ZM184.658 7.46651V13.8548H187.613V7.46651H184.658ZM193.701 4.50335V9.96767H195.853V11.6563H193.701V19.0005H191.645V4.50335H193.701Z"
-      fill="white"
-    />
-    <path
-      d="M208.622 4.50335V19.0324H206.55V4.50335H208.622ZM200.687 5.52293C202.888 5.52293 204.526 7.53023 204.526 10.6846C204.526 13.8707 202.888 15.8621 200.687 15.8621C198.439 15.8621 196.801 13.8707 196.817 10.6846C196.801 7.53023 198.439 5.52293 200.687 5.52293ZM200.687 7.35499C199.547 7.35499 198.792 8.53388 198.792 10.6846C198.792 12.8512 199.547 14.0141 200.687 14.0141C201.796 14.0141 202.55 12.8512 202.55 10.6846C202.55 8.53388 201.796 7.35499 200.687 7.35499Z"
-      fill="white"
-    />
-    <path
-      d="M218.291 6.41507C218.291 8.53388 220.186 10.812 223.495 11.3696L222.595 13.0583C220.049 12.5564 218.154 11.1386 217.215 9.26671C216.243 11.1227 214.348 12.5485 211.818 13.0583L210.935 11.3696C214.179 10.7961 216.09 8.51795 216.106 6.41507V5.36362H218.291V6.41507ZM223.976 15.5435V17.2162H210.597V15.5435H223.976Z"
-      fill="white"
-    />
-    <path
-      d="M237.147 4.50335V19.0324H235.123V11.2581H232.553V9.61719H235.123V4.50335H237.147ZM232.312 5.7141V7.35499H227.526V9.76056H231.782V11.3537H227.526V14.0938C229.55 14.0779 231.172 14.0062 232.987 13.7274L233.196 15.3523C231.06 15.6709 229.148 15.7506 226.643 15.7506H225.486V5.7141H232.312Z"
-      fill="white"
-    />
-    <path
-      d="M251.072 5.7619V7.13196C251.072 8.94809 251.072 10.9872 250.478 14.0619L248.422 13.8548C248.976 11.1147 249 9.07553 249.016 7.37092H240.407V5.7619H251.072ZM252.357 15.5754V17.2322H239.026V15.5754H243.796V10.589H245.836V15.5754H252.357Z"
-      fill="white"
-    />
-    <path
-      d="M266.603 11.1306V12.7874H253.24V11.1306H266.603ZM265.078 8.50202V10.111H254.863V4.8857H264.997V6.52658H256.902V8.50202H265.078ZM259.89 13.6636C263.054 13.6636 264.965 14.6195 264.981 16.3241C264.965 18.0447 263.054 19.0005 259.89 19.0005C256.694 19.0005 254.766 18.0447 254.766 16.3241C254.766 14.6195 256.694 13.6636 259.89 13.6636ZM259.89 15.2089C257.866 15.2089 256.838 15.5913 256.854 16.3241C256.838 17.0888 257.866 17.4552 259.89 17.4552C261.865 17.4552 262.909 17.0888 262.925 16.3241C262.909 15.5913 261.865 15.2089 259.89 15.2089Z"
-      fill="white"
-    />
-    <path
-      d="M276.128 5.79376V7.40278H267.455V5.79376H270.779V4.4237H272.819V5.79376H276.128ZM271.791 7.86478C273.911 7.86478 275.389 8.94809 275.389 10.5412C275.389 12.1343 273.911 13.2176 271.791 13.2176C269.671 13.2176 268.161 12.1343 268.161 10.5412C268.161 8.94809 269.671 7.86478 271.791 7.86478ZM271.791 9.39415C270.812 9.39415 270.153 9.79243 270.169 10.5412C270.153 11.274 270.812 11.6563 271.791 11.6563C272.771 11.6563 273.446 11.274 273.446 10.5412C273.446 9.79243 272.771 9.39415 271.791 9.39415ZM279.051 4.50335V8.26306H281.027V9.91987H279.051V13.5999H276.995V4.50335H279.051ZM279.051 14.2053V19.0005H276.995V15.8303H269.205V14.2053H279.051Z"
-      fill="white"
-    />
-    <path
-      d="M293.795 5.7619V7.17975C293.779 8.94809 293.779 10.9395 293.249 13.8548L291.242 13.6796C291.723 11.0271 291.74 9.05164 291.74 7.37092H282.986V5.7619H293.795ZM295 15.5754V17.2322H281.669V15.5754H284.512V11.0191H286.552V15.5754H288.158V11.0191H290.182V15.5754H295Z"
-      fill="white"
-    />
-  </svg>
-);
+import ChevronDownIcon from './ChevronDownIcon';
+import GSM from './GSM';
 
 const HomeMeisterSection = () => {
   return (
